requirements.txt
joblib==1.4.2
kaggle==1.6.17
kagglehub==0.3.6
matplotlib==3.6.3
numpy==2.2.2
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==9.0.1
Pillow==11.1.0
Requests==2.32.3
retrievalmodels==0.0.1
scikit_learn==1.2.2
streamlit==1.41.1
streamlit_lottie==0.0.5
tensorflow==2.18.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
